+++ pwmTest/Timer1FastBase.h
@@ -35,7 +35,7 @@ void initialize (uint32_t period);
 void set_period_microseconds_delayed(uint32_t microseconds);
 void increment_period();
 void decrement_period();
-void set_pwm_duty(uint8_t pin, uint32_t duty);
+void set_pwm_duty(uint8_t pin, unsigned long duty);
 void increment_pwm_duty(uint8_t pin);
 void decrement_pwm_duty(uint8_t pin);
 void pwm(uint8_t pin, uint32_t duty, uint32_t microseconds, bool invert);
